+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ### Fixed
 
 - Not being able to escape string interpolation.
+- Not able to use if statements inside node body.
 
 ## [0.4.7] - 2026-05-31