+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to `nbcl` are documented here.
 This changelog follows the [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format,
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [0.5.1] - 2026-06-10
+
+### Fixed
+
+- Context not being returned if error in `evaluate_ast_for_ctx`.
+
 ## [0.5.0] - 2026-06-07
 
 ### Added